restructure routes and pages + add market booths and offers skelletons
