+++ src/components/icons/RocketLaunchIcon.tsx
@@ -1,0 +1,19 @@
+import styles from "./IconDefault.module.scss";
+
+type Props = {
+  class?: string;
+  onClick?: (_event?: MouseEvent) => void;
+};
+
+export function RocketLaunchIcon(props: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 -960 960 960"
+      class={props.class || styles.IconDefault}
+      onClick={(e) => props.onClick?.(e)}
+    >
+      <path d="m202.072-560.043 91.964 39.166q15.79-31.021 33.261-61.148 17.471-30.127 37.29-58.454l-68.464-13.616-94.051 94.052Zm152.71 85.094L479.616-350.21q53.435-24.181 101.732-56.264 48.297-32.084 82.725-66.511 76.297-76.457 114.558-160.069 38.26-83.612 39.391-180.301-96.689 1.13-180.428 39.391-83.739 38.261-160.037 114.558-34.427 34.427-66.51 82.725-32.083 48.297-56.265 101.732Zm214.392-89.653q-20.681-20.681-20.681-50.877 0-30.195 20.681-50.876 20.682-20.522 50.924-20.522 30.243 0 50.924 20.522 20.522 20.681 20.522 50.876 0 30.196-20.522 50.877-20.681 20.522-50.924 20.522-30.242 0-50.924-20.522Zm-4.703 367.197 94.29-94.051-13.616-68.464q-28.326 19.819-58.453 37.243-30.127 17.424-61.149 33.054l38.928 92.218Zm327.486-689.979q11.775 138.971-31.656 253.917-43.431 114.945-143.699 215.373-1.145.985-2.33 2.29-1.185 1.304-2.171 2.29l21.334 106.891q3.811 18.978-1.906 36.877-5.717 17.898-19.391 31.572L539.188-64.558l-88.891-207.551L276.775-445.63l-207.55-88.892 173.521-172.949q13.674-13.674 31.62-19.391 17.945-5.718 36.923-1.906l106.892 21.333q.986-.985 2.17-1.677 1.185-.692 2.33-1.678 100.428-100.587 215.366-144.478 114.938-43.891 253.91-32.116ZM143.456-322.63q37.073-37.073 89.924-37.493 52.852-.421 89.765 36.493 36.753 36.753 36.333 89.684-.42 52.932-37.333 89.845-27.421 27.42-87.16 45.83-59.739 18.409-168.572 32.119 13.71-108.834 31.746-169.026 18.036-60.192 45.297-87.452Zm53.616 53.789q-12.348 13.015-22.138 41.029-9.79 28.015-16.456 68.834 40.819-6.667 68.833-16.529 28.015-9.862 40.87-22.21 15.87-14.377 16.297-35.649.428-21.272-14.616-37.142-15.869-15.043-37.141-14.543t-35.649 16.21Z" />
+    </svg>
+  );
+}
