+++ src/components/icons/SearchGlobalIcon.tsx
@@ -1,0 +1,19 @@
+import styles from "./IconDefault.module.scss";
+
+type Props = {
+  class?: string;
+  onClick?: (_event?: MouseEvent) => void;
+};
+
+export function SearchGlobalIcon(props: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 -960 960 960"
+      class={props.class || styles.IconDefault}
+      onClick={(e) => props.onClick?.(e)}
+    >
+      <path d="M459.195-74.29q-80.572-3.782-151.025-37.199T185.677-199.37q-52.04-54.463-82.166-126.56Q73.384-398.027 73.384-480q0-84.333 32.01-158.505 32.009-74.172 87.054-129.163 55.045-54.991 129.239-87.089t158.552-32.098q150.71 0 262.354 95.097 111.643 95.097 137.675 238.519h-77.086q-17.635-79.528-69.717-142.014-52.081-62.486-129.356-93.929v17.015q0 34.037-23.857 58.993-23.857 24.956-57.874 24.956h-83.9v84.038q0 16.708-12.833 27.937-12.834 11.228-29.825 11.228h-81.211v84.537h98.725v122.456h-57.127L157.963-556.348q-4.333 19.254-6.5 38.126-2.166 18.872-2.166 37.744 0 131.435 88.703 226.945 88.702 95.511 221.195 103.33v75.913Zm390.081-22.406-132.24-132.478q-20.761 13.442-44.649 21.043-23.889 7.602-49.148 7.602-73.512 0-125.162-51.535-51.649-51.535-51.649-125.032 0-73.498 51.61-124.867 51.61-51.37 125.107-51.37t124.872 51.5q51.374 51.5 51.374 125.072 0 25.406-7.946 49.261-7.946 23.855-20.699 45.283l132.319 131.572-53.789 53.95ZM622.862-276.282q42.119 0 71.368-29.196 29.248-29.196 29.248-71.479 0-42.183-29.127-71.403-29.127-29.22-71.311-29.22-42.117 0-71.408 29.127-29.291 29.128-29.291 71.312t29.151 71.521q29.151 29.338 71.37 29.338Z" />
+    </svg>
+  );
+}
