+++ src/components/icons/SearchIcon.tsx
@@ -1,0 +1,19 @@
+import styles from "./IconDefault.module.scss";
+
+type Props = {
+  class?: string;
+  onClick?: (_event?: MouseEvent) => void;
+};
+
+export function SearchIcon(props: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 -960 960 960"
+      class={props.class || styles.IconDefault}
+      onClick={(e) => props.onClick?.(e)}
+    >
+      <path d="M791.681-113.174 530.833-373.862q-29.761 24.524-69.28 38.464-39.52 13.941-84.481 13.941-110.274 0-186.742-76.521-76.468-76.522-76.468-185.355 0-108.834 76.522-185.355 76.522-76.522 185.688-76.522 109.167 0 185.355 76.522 76.189 76.521 76.189 185.455 0 43.711-13.602 83.019-13.601 39.308-39.471 72.323L846.138-167.55l-54.457 54.376ZM376.462-397.21q77.437 0 131.419-54.438 53.981-54.439 53.981-131.685 0-77.247-54.005-131.685-54.005-54.439-131.395-54.439-77.945 0-132.396 54.439-54.45 54.438-54.45 131.685 0 77.246 54.427 131.685 54.427 54.438 132.419 54.438Z" />
+    </svg>
+  );
+}
