+++ src/components/icons/StorefrontIcon.tsx
@@ -1,0 +1,19 @@
+import styles from "./IconDefault.module.scss";
+
+type Props = {
+  class?: string;
+  onClick?: (_event?: MouseEvent) => void;
+};
+
+export function StoreFrontIcon(props: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 -960 960 960"
+      class={props.class || styles.IconDefault}
+      onClick={(e) => props.onClick?.(e)}
+    >
+      <path d="M847.986-517.551v330.885q0 30.993-22.427 53.373t-53.486 22.38H187.927q-30.993 0-53.373-22.38t-22.38-53.373v-329.848q-26.732-24.595-36.142-59.973-9.409-35.379 1.924-72.013l42.667-135.333q8.717-28.74 30.558-45.881 21.841-17.141 50.486-17.141h554q29.783 0 52.341 17.268 22.557 17.268 31.369 45.754L882.71-648.5q11.334 36.333 1.392 71.17t-36.116 59.779Zm-278.319-38.173q27.133 0 44.537-17.917 17.405-17.917 13.738-42.707l-23.761-154.594h-88.123V-614q0 23.57 15.302 40.923 15.303 17.353 38.307 17.353Zm-184.742 0q25.144 0 42.414-17.02 17.27-17.019 17.27-41.256v-156.942h-88.124l-23.761 154.594q-3.76 24.138 12.077 42.381t40.124 18.243Zm-180.555 0q21.166 0 36.655-14.935 15.49-14.935 18.25-36.486l24.348-163.797H195.5l-43.442 140.166q-8.428 26.81 6.942 50.931 15.369 24.121 45.37 24.121Zm551.297 0q29.964 0 45.793-23.805 15.83-23.804 7.149-51.247L764.5-770.942h-87.297l24.188 163.797q2.761 21.551 18.234 36.486 15.474 14.935 36.042 14.935Zm-567.74 369.058h584.146v-294.399q-.203 1.014-5.544 1.054-5.34.04-10.884.04-24.29 0-46.442-9.482T665-519.471q-16.985 18.536-40.689 29.018t-51.268 10.482q-28.565 0-50.811-8.562-22.246-8.561-41.899-26.938-16.101 17.194-38.876 26.347-22.776 9.153-50.448 9.153-29.784 0-53.68-10.153-23.895-10.153-41.662-29.347-23.798 20.362-46.482 29.931-22.685 9.569-44.852 9.569-5.014 0-9.703-.04-4.688-.04-6.703-1.054v294.399Zm584.146 0H187.927 772.073Z" />
+    </svg>
+  );
+}
