+++ src/components/layout/index.ts
@@ -1,0 +1,3 @@
+export * from "./Dialog";
+export * from "./Page";
+export * from "./Section";
