+++ src/routes/commerce/MarketBooths.tsx
@@ -1,0 +1,12 @@
+import { Page, Section } from "../../components/layout";
+import styles from "./MarketBooths.module.scss";
+
+export default function MarketBooths() {
+  return (
+    <Page>
+      <Section>
+        <div class={styles.MarketBooths}>Market Booths</div>
+      </Section>
+    </Page>
+  );
+}
