+++ src/routes/dashboard/Offers.module.scss
@@ -1,0 +1,35 @@
+.Title {
+  display: block;
+  margin: 0 auto;
+  padding: 1rem 0;
+  font: var(--font-title);
+  text-align: center;
+}
+
+.Label {
+  display: block;
+  font: var(--font-label);
+}
+
+.Body {
+  font: var(--font-body);
+}
+
+.Details {
+  display: block;
+  font: var(--font-body);
+}
+
+.Actions {
+  display: flex;
+  justify-content: end;
+  align-items: center;
+  gap: 1rem;
+}
+
+.EditSection {
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  gap: 2rem;
+}
