+++ src/routes/info/GetStarted.tsx
@@ -1,9 +1,9 @@
 import { Trans } from "@mbarzda/solid-i18next";
 import { onMount } from "solid-js";
 
-import { DASHBOARD_PATH } from "../App";
-import { authGuardRedirect, buildAuthorizationRequest } from "../lib/auth";
-import { TKEYS } from "../locales/dev";
+import { DASHBOARD_PATH } from "../../App";
+import { authGuardRedirect, buildAuthorizationRequest } from "../../lib/auth";
+import { TKEYS } from "../../locales/dev";
 import styles from "./GetStarted.module.scss";
 
 export default function GetStarted() {
