+++ src/routes/info/PrivacyPolicy.tsx
@@ -1,8 +1,8 @@
 import { Trans } from "@mbarzda/solid-i18next";
 
-import { Page } from "../components/layout/Page";
-import { TKEYS } from "../locales/dev";
-import { Section } from "../components/layout/Section";
+import { Page } from "../../components/layout/Page";
+import { Section } from "../../components/layout/Section";
+import { TKEYS } from "../../locales/dev";
 
 export default function PrivacyPolicy() {
   return (
